+++ client/src/docs/global-docs.tsx
@@ -9,19 +9,41 @@ import { CardBody, CardText } from "reactstrap"
 const THE_BLIND_WATCHMAKER = "https://en.wikipedia.org/wiki/The_Blind_Watchmaker"
 const CHARLES_DARWIN = "https://en.wikipedia.org/wiki/Charles_Darwin"
 const RICHARD_DAWKINS = "https://en.wikipedia.org/wiki/Richard_Dawkins"
+const GITHUB = "https://github.com/beautiful-code-bv/galapagotchi"
+const GITTER = "https://gitter.im/galapagotchis/community"
+const WEBASSEMBLY = "https://webassembly.org/"
+const WEBGL = "https://en.wikipedia.org/wiki/WebGL"
+const THREEJS = "https://threejs.org/"
+const DARWINS_DANGEROUS_IDEA = "https://en.wikipedia.org/wiki/Darwin%27s_Dangerous_Idea"
+const GMAIL = "galapagotchi@gmail.com"
+const TWITTER = "https://twitter.com/galapagotchi"
 
 export enum GlobalDocTitle {
-    Navigation = "Navigation",
-    Hexalot = "Hexalot",
-    Genes = "Genes",
-    Evolution = "Evolution",
-    Mutation = "Mutation",
     Darwin = "Darwin",
+    Developers = "Developers",
+    Evolution = "Evolution",
+    Genes = "Genes",
+    Hexalot = "Hexalot",
+    Island = "Island",
+    Mutation = "Mutation",
+    Navigation = "Navigation",
+    Release = "What stage is it at?",
+    Twitter = "Twitter",
+    Why = "What is this about?",
 }
 
 export interface IGlobalDoc {
     title: GlobalDocTitle
     body: JSX.Element
+}
+
+interface IRef {
+    href: string
+    text: string
+}
+
+function Ref(props: IRef): JSX.Element {
+    return <span> <a href={props.href} target="_blank">{props.text}</a> </span>
 }
 
 export const GLOBAL_DOCS: IGlobalDoc[] = [
@@ -128,4 +150,86 @@ export const GLOBAL_DOCS: IGlobalDoc[] = [
             </CardBody>,
     }
     ,
+    {
+        title: GlobalDocTitle.Why,
+        body:
+            <CardBody>
+                <CardText>
+                    This project aims to give you a fascinating experience that shows you
+                    how things can evolve by Darwin's natural selection.
+                    You are taken to an imaginary island where you can see inside of a toy world,
+                    with toy physics, and you can evolve little robots to run.
+                </CardText>
+            </CardBody>,
+    },
+    {
+        title: GlobalDocTitle.Release,
+        body:
+            <CardBody>
+                This is a very early alpha release of the Galapagotchi universe,
+                but if you want to join in, create your own evolving Galapagotchi and start evolving,
+                contact us and we will send you a key.
+            </CardBody>,
+    },
+    {
+        title: GlobalDocTitle.Island,
+        body:
+            <CardBody>
+                <CardText>
+                    You can visit the island and claim your new hexalot
+                    but remember, every new hexalot creates a little bit more virgin territory.
+                </CardText>
+                <CardText>
+                    It's your job to terraform the new part of the island before you can claim a home there.
+                    Terraforming is done with hexagonal spots of land and water.
+                    It has to follow certain rules:
+                </CardText>
+                <CardText>
+                    There always has to be a land path to the edge of the island,
+                    and land spot needs to touch two land spots.
+                    Also, neither water nor land can be surrounded by its own kind.
+                    That's all.
+                </CardText>
+            </CardBody>,
+    },
+    {
+        title: GlobalDocTitle.Developers,
+        body:
+            <CardBody>
+                <CardText>
+                    <p>
+                        You've come to the right place if you are interested in
+                        <br/>
+                        <br/>
+                        <Ref href={WEBASSEMBLY} text="WebAssembly"/>
+                        <br/>
+                        <Ref href={THREEJS} text="ThreeJS"/>
+                        <br/>
+                        <Ref href={WEBGL} text="WebGL"/>
+                        <br/>
+                        <Ref href={DARWINS_DANGEROUS_IDEA} text="Darwin's Evolution"/>
+                    </p>
+                    <p>
+                        Join
+
+                        <Ref href={GITTER} text="the gitter community"/>
+
+                        or explore
+
+                        <Ref href={GITHUB} text="the github project"/>.
+                    </p>
+
+                </CardText>
+            </CardBody>,
+    },
+    {
+        title: GlobalDocTitle.Twitter,
+        body:
+            <CardBody>
+                <CardText>
+                    Follow the project on Twitter <Ref href={TWITTER} text="@galapagotchi"/> or
+                    contact us at <strong>{GMAIL}</strong>.
+                </CardText>
+            </CardBody>,
+    },
 ]
